updated visibility toggle and moved to common
