+++ components/activity/editor/d2l-activity-editor-footer.js
@@ -1,5 +1,5 @@
 import '@brightspace-ui/core/components/button/button.js';
-import '../visibility/d2l-activity-visibility.js';
+import '../../common/d2l-activity-visibility.js';
 import { css, LitElement } from 'lit-element/lit-element.js';
 import { html } from '@brightspace-hmc/foundation-engine/framework/lit/hypermedia-components.js';
 import { HypermediaStateMixin } from '@brightspace-hmc/foundation-engine/framework/lit/HypermediaStateMixin.js';
@@ -26,7 +26,7 @@ class ActivityEditorFooter extends LocalizeFoundationEditor(HypermediaStateMixin
 			<div>
 				<d2l-button primary @click="${this._onSaveClick}">${this.localize('action.saveClose')}</d2l-button>
 				<d2l-button @click="${this._onCancelClick}">${this.localize('action.cancel')}</d2l-button>
-				<d2l-hc-visibility  href="${this.href}" .token="${this.token}" can-edit-draft></d2l-hc-visibility>
+				<d2l-hc-visibility-toggle  href="${this.href}" .token="${this.token}" can-edit-draft></d2l-hc-visibility-toggle>
 			</div>
 			<div><slot name="save-status">${this.localize('text.saveStatus')}</slot></div>
 		`;
