--- components/activity/visibility/lang/en.js
@@ -1,6 +1,0 @@
-/* eslint quotes: 0 */
-
-export default {
-	"label.hidden": "Hidden",
-	"label.visible": "Visible"
-};
