--- components/activity/visibility/lang/localization.js
@@ -1,9 +1,0 @@
-import { getLocalizeResources } from '../../../../lang/localizeResources.js';
-import { LocalizeMixin } from '@brightspace-ui/core/mixins/localize-mixin.js';
-
-export const LocalizeFoundationVisibility = superclass => class extends LocalizeMixin(superclass) {
-
-	static async getLocalizeResources(langs) {
-		return getLocalizeResources(langs, 'components/activity/description');
-	}
-};
